+++ package/rhythmicon-vue/CHANGES.md
@@ -1,12 +1,10 @@
 # 0.3.0
 
 - RhythmCircle: show pulse index on hover, show symmetry
-- RhythmInput: allow more characters
+- RhythmInput: allow more characters and (reversed) Tracy numbers
 - Change property `rhythm` to Vue model
 - Add RhythmControls component
 
 # 0.2.0
 
 - Add RhythmScore and RhythmSequencer
-
-